requirements.txt
psycopg2==2.9.4
SQLAlchemy==1.4.42
python-dotenv==0.21.0
python-telegram-bot==13.14